[Refactor] 폴링용 Recording 상태 조회 코드를 RecordingQueryService로 이동
