+++ src/main/java/com/dgu/review/domain/interview/controller/RecordingController.java
@@ -3,7 +3,8 @@ package com.dgu.review.domain.interview.controller;
 import com.dgu.review.domain.interview.dto.request.RecordingCreateRequest;
 import com.dgu.review.domain.interview.dto.response.RecordingCreateResponse;
 import com.dgu.review.domain.interview.dto.response.RecordingResultsResponse;
-import com.dgu.review.domain.interview.service.InterviewSessionService;
+import com.dgu.review.domain.interview.service.RecordingCommandService;
+import com.dgu.review.domain.interview.service.RecordingQueryService;
 import com.dgu.review.domain.interview.service.SttService;
 import com.dgu.review.global.response.ApiResponse;
 import jakarta.validation.Valid;
@@ -21,21 +22,22 @@ import org.springframework.web.bind.annotation.*;
 @Validated
 public class RecordingController {
 
-    private final InterviewSessionService interviewSessionService;
+    private final RecordingCommandService recordingCommandService;
     private final SttService sttService;
+    private final RecordingQueryService recordingQueryService;
 
     @PostMapping("/questions/{questionId}/recordings")
     public ResponseEntity<ApiResponse<RecordingCreateResponse>> submitRecording(
             @PathVariable Long questionId,
             @Valid @RequestBody RecordingCreateRequest request
     ) {
-        RecordingCreateResponse response = interviewSessionService.createAndTranscribe(questionId, request);
+        RecordingCreateResponse response = recordingCommandService.createAndTranscribe(questionId, request);
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(ApiResponse.ok(response));
     }
 
     @GetMapping("/recordings/{recordingId}/results")
     public ResponseEntity<ApiResponse<RecordingResultsResponse>> getRecordingStatus(@PathVariable Long recordingId) {
-        var res = sttService.getRecordingResults(recordingId);
+        var res = recordingQueryService.getRecordingResults(recordingId);
         return ResponseEntity.ok(ApiResponse.ok(res));
     }
