+++ src/main/java/com/dgu/review/domain/interview/service/RecordingCommandService.java
@@ -18,7 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class RecordingService {
+public class RecordingCommandService {
 
     private final RecordingRepository recordingRepo;
     private final RecordingStatusService statusService;
